fix subtle bug related to random numbers.
random.random is used for snake food, body position random generation.
numpy.random is used for snake brain (DNN) generation and genetic
algorithm optimization.
Separate these two things makes me able to call random.seed() to control
the environment of the game (the food generation), while the randomness
of snake's brain and genetic algorithm is not effected.

fix a bug related to a state flag of snake. Ideally, with the same
seed for random.seed(), the environment is the same for the snake,
and the snake should give the same results after multiple run.
However, because one state flag is not reset right, the repeating
behaviour is not achieve. This is the bug and now it is fixed.